+++ crates/github-actions-models/src/workflow/job.rs
@@ -106,6 +106,11 @@ pub struct Step {
     #[serde(default)]
     pub continue_on_error: BoE,
 
+    /// Whether the step runs asynchronously, i.e. does not block its successor from running.
+    // TODO: Is this allowed to be an expression?
+    #[serde(default)]
+    pub background: bool,
+
     /// An optional environment mapping for this step.
     #[serde(default)]
     pub env: LoE<Env>,
@@ -116,29 +121,61 @@ pub struct Step {
 }
 
 #[derive(Deserialize, Debug)]
-#[serde(rename_all = "kebab-case", untagged)]
+#[serde(rename_all_fields = "kebab-case", untagged)]
 pub enum StepBody {
-    Uses {
-        /// The GitHub Action being used.
-        #[serde(deserialize_with = "crate::common::step_uses")]
-        uses: Uses,
-
-        /// Any inputs to the action being used.
-        #[serde(default)]
-        with: LoE<Env>,
+    Uses(UsesBody),
+    Run(RunBody),
+    Wait {
+        /// One or more steps, by ID, that this step is blocked by (i.e. waits for).
+        // TODO: Is this allowed to be an expression?
+        #[serde(deserialize_with = "crate::common::scalar_or_vector")]
+        wait: Vec<String>,
     },
-    Run {
-        /// The command to run.
-        #[serde(deserialize_with = "crate::common::bool_is_string")]
-        run: String,
-
-        /// An optional working directory to run [`StepBody::Run::run`] from.
-        working_directory: Option<String>,
-
-        /// An optional shell to run in. Defaults to the job or workflow's
-        /// default shell.
-        shell: Option<LoE<String>>,
+    WaitAll {
+        /// A marker indicating that this step waits for all active background steps.
+        #[serde(deserialize_with = "crate::common::bool_or_unit")]
+        wait_all: bool,
     },
+    Cancel {
+        /// A background step, by ID, that this step terminates.
+        cancel: String,
+    },
+    Parallel {
+        /// One or more steps to run in parallel.
+        // NOTE: This is an overapproximation, since parallel blocks
+        // do not allow nested `parallel` locks, or pseudo-steps like `cancel`.
+        // TODO: Ratchet this down to just the steps that parallel blocks do allow.
+        parallel: Vec<Step>,
+    },
+}
+
+/// The body of a `uses:`-style step clause.
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "kebab-case")]
+pub struct UsesBody {
+    /// The GitHub Action being used.
+    #[serde(deserialize_with = "crate::common::step_uses")]
+    pub uses: Uses,
+
+    /// Any inputs to the action being used.
+    #[serde(default)]
+    pub with: LoE<Env>,
+}
+
+/// The body of a `run:`-style step clause.
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "kebab-case")]
+pub struct RunBody {
+    /// The command to run.
+    #[serde(deserialize_with = "crate::common::bool_is_string")]
+    pub run: String,
+
+    /// An optional working directory to run [`StepBody::Run::run`] from.
+    pub working_directory: Option<String>,
+
+    /// An optional shell to run in. Defaults to the job or workflow's
+    /// default shell.
+    pub shell: Option<LoE<String>>,
 }
 
 #[derive(Deserialize, Debug)]
@@ -210,7 +247,7 @@ pub enum Secrets {
 mod tests {
     use crate::{
         common::{EnvValue, expr::LoE},
-        workflow::job::{Matrix, Secrets},
+        workflow::job::{Matrix, Secrets, StepBody},
     };
 
     use super::{RunsOn, Strategy};
@@ -273,5 +310,25 @@ matrix:
                 .to_string(),
             "runs-on must provide either `group` or one or more `labels`"
         );
+    }
+
+    #[test]
+    fn test_stepbody_working_directory() {
+        let step = r#"
+run: foo
+working-directory: /tmp
+"#;
+
+        insta::assert_debug_snapshot!(&yaml_serde::from_str::<StepBody>(step).unwrap(), @r#"
+        Run(
+            RunBody {
+                run: "foo",
+                working_directory: Some(
+                    "/tmp",
+                ),
+                shell: None,
+            },
+        )
+        "#);
     }
 }