+++ crates/zizmor/src/audit/github_env.rs
@@ -2,7 +2,7 @@ use std::ops::{Deref as _, Range};
 
 use anyhow::{Context as _, Result};
 use github_actions_models::action;
-use github_actions_models::workflow::job::StepBody;
+use github_actions_models::workflow::job::{RunBody, StepBody};
 use tree_sitter::{
     Language, Parser, QueryCapture, QueryCursor, QueryMatches, StreamingIterator as _, Tree,
 };
@@ -13,7 +13,7 @@ use crate::config::Config;
 use crate::finding::location::Locatable as _;
 use crate::finding::{Confidence, Finding, Severity};
 use crate::models::StepCommon as _;
-use crate::models::{workflow::JobCommon as _, workflow::Step};
+use crate::models::workflow::Step;
 use crate::state::AuditState;
 use crate::utils;
 use crate::utils::once::static_regex;
@@ -393,13 +393,12 @@ impl Audit for GitHubEnv {
             return Ok(findings);
         }
 
-        if let StepBody::Run { run, .. } = &step.deref().body {
+        if let StepBody::Run(RunBody { run, .. }) = &step.deref().body {
             let shell = step.shell().map(|s| s.0).unwrap_or_else(|| {
                 tracing::warn!(
-                    "github-env: couldn't determine shell type for {workflow}:{job} step {stepno}; assuming bash",
+                    "github-env: couldn't determine shell type for {workflow} step {loc:#?}; assuming bash",
                     workflow = step.workflow().key.presentation_path(),
-                    job = step.parent.id(),
-                    stepno = step.index
+                    loc = step.location(),
                 );
 
                 // If we can't infer a shell for this `run:`, assume that it's